Delete helpers and duplicate prometheus tests
